+++ dropwizard-auth/src/test/java/io/dropwizard/auth/basic/BasicAuthProviderTest.java
@@ -3,10 +3,11 @@ package io.dropwizard.auth.basic;
 import com.codahale.metrics.MetricRegistry;
 import com.google.common.base.Optional;
 import io.dropwizard.auth.Auth;
-import io.dropwizard.auth.AuthFactory;
+import io.dropwizard.auth.AuthDynamicFeature;
 import io.dropwizard.auth.AuthResource;
 import io.dropwizard.auth.AuthenticationException;
 import io.dropwizard.auth.Authenticator;
+import io.dropwizard.auth.PrincipalImpl;
 import io.dropwizard.jersey.DropwizardResourceConfig;
 import io.dropwizard.logging.LoggingFactory;
 import org.glassfish.jersey.servlet.ServletProperties;
@@ -17,17 +18,18 @@ import org.glassfish.jersey.test.grizzly.GrizzlyWebTestContainerFactory;
 import org.glassfish.jersey.test.spi.TestContainerException;
 import org.glassfish.jersey.test.spi.TestContainerFactory;
 import org.junit.Test;
-
 import javax.ws.rs.GET;
 import javax.ws.rs.Path;
 import javax.ws.rs.Produces;
 import javax.ws.rs.WebApplicationException;
+import javax.ws.rs.core.Context;
 import javax.ws.rs.core.HttpHeaders;
 import javax.ws.rs.core.MediaType;
+import javax.ws.rs.core.SecurityContext;
+import java.security.Principal;
 
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.assertj.core.api.Assertions.failBecauseExceptionWasNotThrown;
-
 
 public class BasicAuthProviderTest extends JerseyTest {
     static {
@@ -57,6 +59,20 @@ public class BasicAuthProviderTest extends JerseyTest {
             assertThat(e.getResponse().getHeaders().get(HttpHeaders.WWW_AUTHENTICATE))
                     .containsOnly("Basic realm=\"realm\"");
         }
+    }
+
+    @Test
+    public void resourceWithoutAuth200() {
+        assertThat(target("/test/noauth").request()
+                .get(String.class))
+                .isEqualTo("hello");
+    }
+
+    @Test
+    public void resourceWithAuthNotRequired200() {
+        assertThat(target("/test/authnotrequired").request()
+                .get(String.class))
+                .isEqualTo("No Principal");
     }
 
     @Test
@@ -97,9 +113,10 @@ public class BasicAuthProviderTest extends JerseyTest {
     @Path("/test/")
     @Produces(MediaType.TEXT_PLAIN)
     public static class ExampleResource {
+        @Auth
         @GET
-        public String show(@Auth String principal) {
-            return principal;
+        public String show(@Context SecurityContext context) {
+            return context.getUserPrincipal().getName();
         }
     }
 
@@ -107,12 +124,12 @@ public class BasicAuthProviderTest extends JerseyTest {
         public BasicAuthTestResourceConfig() {
             super(true, new MetricRegistry());
 
-            final Authenticator<BasicCredentials, String> authenticator = new Authenticator<BasicCredentials, String>() {
+            final Authenticator<BasicCredentials, Principal> authenticator = new Authenticator<BasicCredentials, Principal>() {
                 @Override
-                public Optional<String> authenticate(BasicCredentials credentials) throws AuthenticationException {
+                public Optional<Principal> authenticate(BasicCredentials credentials) throws AuthenticationException {
                     if ("good-guy".equals(credentials.getUsername()) &&
                             "secret".equals(credentials.getPassword())) {
-                        return Optional.of("good-guy");
+                        return Optional.<Principal>of(new PrincipalImpl("good-guy"));
                     }
                     if ("bad-guy".equals(credentials.getUsername())) {
                         throw new AuthenticationException("CRAP");
@@ -120,7 +137,11 @@ public class BasicAuthProviderTest extends JerseyTest {
                     return Optional.absent();
                 }
             };
-            register(AuthFactory.binder(new BasicAuthFactory<>(authenticator, "realm", String.class)));
+
+            register(new AuthDynamicFeature(
+                    new BasicCredentialAuthHandler.Builder<>()
+                            .setAuthenticator(authenticator)
+                            .buildAuthHandler()));
             register(AuthResource.class);
         }
     }